Add period generating views
These view contain queries that ensure that all periods for each
itemsite or location/lot/serial are included from the earliest time the
item is transacted to last time (or to the present if there are current
quantities on hand).

The queries ensure that any periods of dormancy are represented in the
list of periods that need to be part of calculations.  This should help
prune any periods that are before the item makes an active appearance
and allow them to disappear from the time line once they cease having
transactions or current quantities on hand.