Add real-time history reporting function
These are the original queries that were written to replace the as-of
reporting functionality.  These should work fine, though they weren't
suitable for the sponsoring client as they are using a tool that can't
easily query table returning database functions.

Still, they are presented here as they don't need refreshing from as do
the summary tables and they are a good double-check for reporting
accuracy.